requirements.txt
PyQt5>=5.15.0
PyYAML>=5.4.0

